fix(manipulation): encapsulate VAMP planner state
